+++ simpleble/src/backends/macos/PeripheralBaseMacOS.mm
@@ -83,7 +83,7 @@ typedef struct {
         }
 
         if (self.peripheral.state != CBPeripheralStateConnected) {
-            throw SimpleBLE::Exception::OperationFailed();
+            throw SimpleBLE::Exception::OperationFailed("Peripheral Connection");
         }
 
         [self.peripheral discoverServices:nil];
@@ -98,7 +98,7 @@ typedef struct {
         if (self.peripheral.services == nil) {
             // If services could not be discovered, raise an exception.
             NSLog(@"Services could not be discovered.");
-            throw SimpleBLE::Exception::OperationFailed();
+            throw SimpleBLE::Exception::OperationFailed("Service Discovery");
         }
 
         // For each service found, discover characteristics.
@@ -115,7 +115,7 @@ typedef struct {
             if (service.characteristics == nil) {
                 // If characteristics could not be discovered, raise an exception.
                 NSLog(@"Characteristics could not be discovered for service %@", service.UUID);
-                throw SimpleBLE::Exception::OperationFailed();
+                throw SimpleBLE::Exception::OperationFailed("Characteristic Discovery");
             }
 
             // For each characteristic, create the associated extra properties and discover descriptors.
@@ -131,7 +131,7 @@ typedef struct {
                 if (characteristic.descriptors == nil) {
                     // If characteristics could not be discovered, raise an exception.
                     NSLog(@"Descriptors could not be discovered for characteristic %@", characteristic.UUID);
-                    throw SimpleBLE::Exception::OperationFailed();
+                    throw SimpleBLE::Exception::OperationFailed("Descriptor Discovery");
                 }
 
                 characteristic_extras_t characteristic_extra;
@@ -167,7 +167,7 @@ typedef struct {
         if (self.peripheral.state != CBPeripheralStateDisconnected) {
             // If the disconnection failed, raise an exception.
             NSLog(@"Disconnection failed.");
-            throw SimpleBLE::Exception::OperationFailed();
+            throw SimpleBLE::Exception::OperationFailed("Peripheral Disconnection");
         }
     }
 }
@@ -232,7 +232,7 @@ typedef struct {
 
     if (readPending) {
         NSLog(@"Characteristic %@ could not be read", characteristic.UUID);
-        throw SimpleBLE::Exception::OperationFailed();
+        throw SimpleBLE::Exception::OperationFailed("Characteristic Read");
     }
 
     return SimpleBLE::ByteArray((const char*)characteristic.value.bytes, characteristic.value.length);
@@ -267,7 +267,7 @@ typedef struct {
 
     if (writePending) {
         NSLog(@"Characteristic %@ could not be written", characteristic.UUID);
-        throw SimpleBLE::Exception::OperationFailed();
+        throw SimpleBLE::Exception::OperationFailed("Characteristic Write Request");
     }
 }
 
@@ -310,7 +310,7 @@ typedef struct {
 
     if (!characteristic.isNotifying) {
         NSLog(@"Could not enable notifications for characteristic %@", characteristic.UUID);
-        throw SimpleBLE::Exception::OperationFailed();
+        throw SimpleBLE::Exception::OperationFailed("Characteristic Notify/Indicate");
     }
 }
 
@@ -336,7 +336,7 @@ typedef struct {
 
         if (characteristic.isNotifying) {
             NSLog(@"Could not disable notifications for characteristic %@", characteristic.UUID);
-            throw SimpleBLE::Exception::OperationFailed();
+            throw SimpleBLE::Exception::OperationFailed("Characteristic Unsubscribe");
         } else {
             // Only delete the callback if the characteristic is no longer notifying, to
             // prevent triggering a segfault.
@@ -374,7 +374,7 @@ typedef struct {
 
     if (readPending) {
         NSLog(@"Descriptor %@ could not be read", descriptor.UUID);
-        throw SimpleBLE::Exception::OperationFailed();
+        throw SimpleBLE::Exception::OperationFailed("Descriptor Read");
     }
 
     const char* bytes = (const char*)[descriptor.value bytes];
@@ -412,7 +412,7 @@ typedef struct {
 
     if (writePending) {
         NSLog(@"Descriptor %@ could not be written", descriptor.UUID);
-        throw SimpleBLE::Exception::OperationFailed();
+        throw SimpleBLE::Exception::OperationFailed("Descriptor Write");
     }
 }